fix array sort self-hosting by moving predicate resolution to array.ts
