+++ src/codegen/types/collections/array.ts
@@ -24,7 +24,11 @@ interface CallExpr {
 import { IGeneratorContext } from "../../infrastructure/generator-context.js";
 import { generateArrayReverse, generateArrayShift, generateArrayUnshift } from "./array/reorder.js";
 import { generateArrayIndexOf, generateArrayFindIndex } from "./array/search.js";
-import { generateArraySort } from "./array/sort.js";
+import {
+  generateDefaultNumericSort,
+  generateDefaultStringSort,
+  generateNumericSortWithFn,
+} from "./array/sort.js";
 
 export class ArrayGenerator {
   constructor(private ctx: IGeneratorContext) {}
@@ -3503,6 +3507,43 @@ export class ArrayGenerator {
   }
 
   generateArraySort(expr: MethodCallNode, params: string[]): string {
-    return generateArraySort(this.ctx, expr, params);
+    if (expr.args.length > 1) {
+      throw new Error("sort() expects 0 or 1 arguments");
+    }
+
+    const arrayPtr = this.ctx.generateExpression(expr.object, params);
+    this.ctx.setUsesArraySort(true);
+
+    let isStringArray = false;
+    const exprObjBase = expr.object as ExprBase;
+    if (exprObjBase.type === "variable") {
+      const varName = (expr.object as VariableNode).name;
+      const varType = this.ctx.getVariableType(varName);
+      isStringArray = varType === "%StringArray*" || varType === "%StringArray";
+    }
+    if (!isStringArray) {
+      const ptrType = this.ctx.getVariableType(arrayPtr);
+      if (ptrType === "%StringArray*" || ptrType === "%StringArray") isStringArray = true;
+    }
+
+    if (expr.args.length === 0) {
+      if (isStringArray) {
+        return generateDefaultStringSort(this.ctx, arrayPtr);
+      }
+      return generateDefaultNumericSort(this.ctx, arrayPtr);
+    }
+
+    const predicateArg = expr.args[0];
+    let compareFn: string;
+
+    if (predicateArg.type === "variable") {
+      compareFn = this.ctx.mangleUserName((predicateArg as VariableNode).name);
+    } else if (predicateArg.type === "arrow_function") {
+      compareFn = this.ctx.generateExpression(predicateArg, params);
+    } else {
+      throw new Error("sort() comparator must be a function name or inline function");
+    }
+
+    return generateNumericSortWithFn(this.ctx, arrayPtr, compareFn);
   }
 }
