+++ src/codegen/types/collections/array/sort.ts
@@ -1,19 +1,11 @@
-import { Expression, MethodCallNode, VariableNode } from "../../../../ast/types.js";
-
-interface ExprBase {
-  type: string;
-}
+import { Expression } from "../../../../ast/types.js";
 
 interface ArraySortContext {
   nextTemp(): string;
   nextLabel(prefix: string): string;
   emit(instruction: string): void;
-  getVariableType(name: string): string | undefined;
   setVariableType(name: string, type: string): void;
   generateExpression(expr: Expression, params: string[]): string;
-  mangleUserName(name: string): string;
-  setExpectedCallbackParamType(type: string | null): void;
-  setUsesArraySort(value: boolean): void;
 }
 
 export function generateDefaultSortComparators(): string {
@@ -51,41 +43,7 @@ export function generateDefaultSortComparators(): string {
   return ir;
 }
 
-export function generateArraySort(
-  gen: ArraySortContext,
-  expr: MethodCallNode,
-  params: string[],
-): string {
-  if (expr.args.length > 1) {
-    throw new Error("sort() expects 0 or 1 arguments");
-  }
-
-  const arrayPtr = gen.generateExpression(expr.object, params);
-  gen.setUsesArraySort(true);
-
-  let isStringArray = false;
-  const exprObjBase = expr.object as ExprBase;
-  if (exprObjBase.type === "variable") {
-    const varName = (expr.object as VariableNode).name;
-    const varType = gen.getVariableType(varName);
-    isStringArray = varType === "%StringArray*" || varType === "%StringArray";
-  }
-  if (!isStringArray) {
-    const ptrType = gen.getVariableType(arrayPtr);
-    if (ptrType === "%StringArray*" || ptrType === "%StringArray") isStringArray = true;
-  }
-
-  if (expr.args.length === 1) {
-    return generateSortWithCompareFn(gen, expr, params, arrayPtr, isStringArray);
-  }
-
-  if (isStringArray) {
-    return generateDefaultStringSort(gen, arrayPtr);
-  }
-  return generateDefaultNumericSort(gen, arrayPtr);
-}
-
-function generateDefaultNumericSort(gen: ArraySortContext, arrayPtr: string): string {
+export function generateDefaultNumericSort(gen: ArraySortContext, arrayPtr: string): string {
   const lenPtr = gen.nextTemp();
   gen.emit(`${lenPtr} = getelementptr inbounds %Array, %Array* ${arrayPtr}, i32 0, i32 1`);
   const length = gen.nextTemp();
@@ -109,7 +67,7 @@ function generateDefaultNumericSort(gen: ArraySortContext, arrayPtr: string): st
   return arrayPtr;
 }
 
-function generateDefaultStringSort(gen: ArraySortContext, arrayPtr: string): string {
+export function generateDefaultStringSort(gen: ArraySortContext, arrayPtr: string): string {
   const lenPtr = gen.nextTemp();
   gen.emit(
     `${lenPtr} = getelementptr inbounds %StringArray, %StringArray* ${arrayPtr}, i32 0, i32 1`,
@@ -137,31 +95,7 @@ function generateDefaultStringSort(gen: ArraySortContext, arrayPtr: string): str
   return arrayPtr;
 }
 
-function generateSortWithCompareFn(
-  gen: ArraySortContext,
-  expr: MethodCallNode,
-  params: string[],
-  arrayPtr: string,
-  isStringArray: boolean,
-): string {
-  const predicateArg = expr.args[0];
-  let compareFn: string;
-
-  if (predicateArg.type === "variable") {
-    compareFn = gen.mangleUserName((predicateArg as VariableNode).name);
-  } else if (predicateArg.type === "arrow_function") {
-    compareFn = gen.generateExpression(predicateArg, params);
-  } else {
-    throw new Error("sort() comparator must be a function name or inline function");
-  }
-
-  if (isStringArray) {
-    return generateStringArraySortWithFn(gen, arrayPtr, compareFn);
-  }
-  return generateNumericArraySortWithFn(gen, arrayPtr, compareFn);
-}
-
-function generateNumericArraySortWithFn(
+export function generateNumericSortWithFn(
   gen: ArraySortContext,
   arrayPtr: string,
   compareFn: string,
@@ -261,112 +195,5 @@ function generateNumericArraySortWithFn(
   gen.emit(`${endLabel}:`);
 
   gen.setVariableType(arrayPtr, "%Array*");
-  return arrayPtr;
-}
-
-function generateStringArraySortWithFn(
-  gen: ArraySortContext,
-  arrayPtr: string,
-  compareFn: string,
-): string {
-  const lenPtr = gen.nextTemp();
-  gen.emit(
-    `${lenPtr} = getelementptr inbounds %StringArray, %StringArray* ${arrayPtr}, i32 0, i32 1`,
-  );
-  const length = gen.nextTemp();
-  gen.emit(`${length} = load i32, i32* ${lenPtr}`);
-
-  const dataPtrField = gen.nextTemp();
-  gen.emit(
-    `${dataPtrField} = getelementptr inbounds %StringArray, %StringArray* ${arrayPtr}, i32 0, i32 0`,
-  );
-  const dataPtr = gen.nextTemp();
-  gen.emit(`${dataPtr} = load i8**, i8*** ${dataPtrField}, !tbaa !5`);
-
-  const checkLabel = gen.nextLabel("sort_check");
-  const outerBody = gen.nextLabel("sort_outer");
-  const innerCheck = gen.nextLabel("sort_inner_check");
-  const innerBody = gen.nextLabel("sort_inner_body");
-  const swapLabel = gen.nextLabel("sort_swap");
-  const noSwapLabel = gen.nextLabel("sort_noswap");
-  const innerNext = gen.nextLabel("sort_inner_next");
-  const outerNext = gen.nextLabel("sort_outer_next");
-  const endLabel = gen.nextLabel("sort_end");
-
-  const iPtr = gen.nextTemp();
-  gen.emit(`${iPtr} = alloca i32`);
-  gen.emit(`store i32 0, i32* ${iPtr}`);
-
-  const lenMinus1 = gen.nextTemp();
-  gen.emit(`${lenMinus1} = sub i32 ${length}, 1`);
-
-  gen.emit(`br label %${checkLabel}`);
-
-  gen.emit(`${checkLabel}:`);
-  const i = gen.nextTemp();
-  gen.emit(`${i} = load i32, i32* ${iPtr}`);
-  const outerCond = gen.nextTemp();
-  gen.emit(`${outerCond} = icmp slt i32 ${i}, ${lenMinus1}`);
-  gen.emit(`br i1 ${outerCond}, label %${outerBody}, label %${endLabel}`);
-
-  gen.emit(`${outerBody}:`);
-  const jPtr = gen.nextTemp();
-  gen.emit(`${jPtr} = alloca i32`);
-  gen.emit(`store i32 0, i32* ${jPtr}`);
-
-  const remaining = gen.nextTemp();
-  gen.emit(`${remaining} = sub i32 ${lenMinus1}, ${i}`);
-
-  gen.emit(`br label %${innerCheck}`);
-
-  gen.emit(`${innerCheck}:`);
-  const j = gen.nextTemp();
-  gen.emit(`${j} = load i32, i32* ${jPtr}`);
-  const innerCond = gen.nextTemp();
-  gen.emit(`${innerCond} = icmp slt i32 ${j}, ${remaining}`);
-  gen.emit(`br i1 ${innerCond}, label %${innerBody}, label %${outerNext}`);
-
-  gen.emit(`${innerBody}:`);
-  const ptrA = gen.nextTemp();
-  gen.emit(`${ptrA} = getelementptr inbounds i8*, i8** ${dataPtr}, i32 ${j}`);
-  const valA = gen.nextTemp();
-  gen.emit(`${valA} = load i8*, i8** ${ptrA}`);
-
-  const jPlus1 = gen.nextTemp();
-  gen.emit(`${jPlus1} = add i32 ${j}, 1`);
-  const ptrB = gen.nextTemp();
-  gen.emit(`${ptrB} = getelementptr inbounds i8*, i8** ${dataPtr}, i32 ${jPlus1}`);
-  const valB = gen.nextTemp();
-  gen.emit(`${valB} = load i8*, i8** ${ptrB}`);
-
-  const cmpResult = gen.nextTemp();
-  gen.emit(`${cmpResult} = call double @${compareFn}(i8* ${valA}, i8* ${valB})`);
-  const shouldSwap = gen.nextTemp();
-  gen.emit(`${shouldSwap} = fcmp ogt double ${cmpResult}, 0.0`);
-  gen.emit(`br i1 ${shouldSwap}, label %${swapLabel}, label %${noSwapLabel}`);
-
-  gen.emit(`${swapLabel}:`);
-  gen.emit(`store i8* ${valB}, i8** ${ptrA}`);
-  gen.emit(`store i8* ${valA}, i8** ${ptrB}`);
-  gen.emit(`br label %${innerNext}`);
-
-  gen.emit(`${noSwapLabel}:`);
-  gen.emit(`br label %${innerNext}`);
-
-  gen.emit(`${innerNext}:`);
-  const nextJ = gen.nextTemp();
-  gen.emit(`${nextJ} = add i32 ${j}, 1`);
-  gen.emit(`store i32 ${nextJ}, i32* ${jPtr}`);
-  gen.emit(`br label %${innerCheck}`);
-
-  gen.emit(`${outerNext}:`);
-  const nextI = gen.nextTemp();
-  gen.emit(`${nextI} = add i32 ${i}, 1`);
-  gen.emit(`store i32 ${nextI}, i32* ${iPtr}`);
-  gen.emit(`br label %${checkLabel}`);
-
-  gen.emit(`${endLabel}:`);
-
-  gen.setVariableType(arrayPtr, "%StringArray*");
   return arrayPtr;
 }
